CLEVER Initial release
CLEVER (Cross-Lipschitz Extreme Value for nEtwork Robustness) is a
metric for measuring the robustness of deep neural networks. It
estimates the robustness lower bound by sampling the norm of gradients
and fitting a limit distribution using extreme value theory. CLEVER
score is attack-agnostic; a higher score number indicates that the
network is likely to be less venerable to adversarial examples. CLEVER
can be efficiently computed even for large state-of-the-art ImageNet
models like ResNet-50 and Inception-v3.