Fix failure in CLEVER score computation (see Issue #7)
In recent version of scipy.io, the behavoir of savemat() function
changed. It does not append the .mat extension to filenames anymore, so
the collected gradients (which are saved to .mat files) are not found by
clever.py. The score calcuation will fail with a weird error.

Additional, the newer version matplotlib also has some problem in
plotting. The "normed" parameter has been removed in hist() and must be
replaced by "density". This has also been fixed in this commit.  I also
cleaned up the logic for plotting, to avoid generating unnecessary
plots.